client/pyproject.toml
[project]
name = "robot-streamer-demo"
version = "0.1.0"
description = "Robot telemetry streaming demo for FortyFive ML Dashboard"
readme = "../README.md"
requires-python = ">=3.11"

dependencies = [
    "websocket-client>=1.7.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]